drop all uses of basename(2)
the POSIX variant of basename(2) can freely write
to the string it is passed. in all call sites, we
were passing a const char* ultimately sourced
from argv[0], to which we must not write. the GNU
variant doesn't have this behavior, but we can't
assume that variant on e.g. musl.

so just drop the basename() calls entirely. they
didn't add much, and arguably lost information.

thanks to znley and celeste of alpine for bringing
this to my attention in

https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/71142

 alpine linux: always kicking ass